// apps/frontend/src/app/api/models/device-telemetry-state-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { JsonNode } from '../models/json-node';
export interface DeviceTelemetryStateDto {
  data?: JsonNode;
  deviceId?: string;
  messageType?: 'TELEMETRY' | 'STATUS' | 'COMMAND' | 'COMMAND_RESULT' | 'HELLO' | 'STATE' | 'ERROR' | 'CONFIG';
  receivedAt?: string;
  reportedAt?: string;
  schemaVersion?: number;
}